# griffin-runner/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "griffin-runner API",
    "description": "API for managing and running test plans",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/runs/runs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "planId",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": ["pending"]
                },
                {
                  "type": "string",
                  "enum": ["running"]
                },
                {
                  "type": "string",
                  "enum": ["completed"]
                },
                {
                  "type": "string",
                  "enum": ["failed"]
                }
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["runs", "total", "limit", "offset"],
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "planId",
                          "planName",
                          "status",
                          "triggeredBy",
                          "startedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "planId": {
                            "type": "string"
                          },
                          "planName": {
                            "type": "string"
                          },
                          "status": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": ["pending"]
                              },
                              {
                                "type": "string",
                                "enum": ["running"]
                              },
                              {
                                "type": "string",
                                "enum": ["completed"]
                              },
                              {
                                "type": "string",
                                "enum": ["failed"]
                              }
                            ]
                          },
                          "triggeredBy": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": ["schedule"]
                              },
                              {
                                "type": "string",
                                "enum": ["manual"]
                              },
                              {
                                "type": "string",
                                "enum": ["api"]
                              }
                            ]
                          },
                          "startedAt": {
                            "type": "string"
                          },
                          "completedAt": {
                            "type": "string"
                          },
                          "duration_ms": {
                            "type": "number"
                          },
                          "success": {
                            "type": "boolean"
                          },
                          "errors": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/runs/runs/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "planId",
                    "planName",
                    "status",
                    "triggeredBy",
                    "startedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "planId": {
                      "type": "string"
                    },
                    "planName": {
                      "type": "string"
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": ["pending"]
                        },
                        {
                          "type": "string",
                          "enum": ["running"]
                        },
                        {
                          "type": "string",
                          "enum": ["completed"]
                        },
                        {
                          "type": "string",
                          "enum": ["failed"]
                        }
                      ]
                    },
                    "triggeredBy": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": ["schedule"]
                        },
                        {
                          "type": "string",
                          "enum": ["manual"]
                        },
                        {
                          "type": "string",
                          "enum": ["api"]
                        }
                      ]
                    },
                    "startedAt": {
                      "type": "string"
                    },
                    "completedAt": {
                      "type": "string"
                    },
                    "duration_ms": {
                      "type": "number"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/runs/plan/{id}/run": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["jobRun", "message"],
                  "properties": {
                    "jobRun": {
                      "type": "object",
                      "required": [
                        "id",
                        "planId",
                        "planName",
                        "status",
                        "triggeredBy",
                        "startedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "planId": {
                          "type": "string"
                        },
                        "planName": {
                          "type": "string"
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": ["pending"]
                            },
                            {
                              "type": "string",
                              "enum": ["running"]
                            },
                            {
                              "type": "string",
                              "enum": ["completed"]
                            },
                            {
                              "type": "string",
                              "enum": ["failed"]
                            }
                          ]
                        },
                        "triggeredBy": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": ["schedule"]
                            },
                            {
                              "type": "string",
                              "enum": ["manual"]
                            },
                            {
                              "type": "string",
                              "enum": ["api"]
                            }
                          ]
                        },
                        "startedAt": {
                          "type": "string"
                        },
                        "completedAt": {
                          "type": "string"
                        },
                        "duration_ms": {
                          "type": "number"
                        },
                        "success": {
                          "type": "boolean"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
}